抽离ASM_SmartRunPluginImp为单独项目

+++ by122006library/src/main/java/com/by122006library/Functions/ViewsReplace.java
@@ -137,7 +137,7 @@ public class ViewsReplace {
             att1.setReverse();
             ViewAtt att2 = new ViewAtt(to, ViewAtt.AttStyle.Alpha, 1, 0);
 //        att2.logProgress("to");
-            time = new TimeAtt((long) (durTime)) {
+            time = new TimeAtt( durTime) {
                 @Override
                 @UIThread
                 public void remove() {
